% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohnAfunctions.R
\name{cheat}
\alias{cheat}
\title{Giving you my answer for an exercise from 3.1}
\usage{
cheat(exercise_number)
}
\arguments{
\item{exercise_number}{Requires to be a number between 1 to 10.}
}
\description{
This function returns my answer for the first 10 questions from the 3.1 exercise sheet.
}
\examples{
cheat(1)
# Q3.1.1

my_sample <- rnorm(1000)

boxplot(my_sample))
}
\keyword{Exercises}
\keyword{PiPS}
\keyword{Q3.1}
\keyword{Q3.2.2}
