% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohnAfunctions.R
\name{remind_me}
\alias{remind_me}
\title{A function to quickly retrieve my current Steam history}
\usage{
remind_me()
}
\arguments{
\item{...}{There are no parameters for this function.}
}
\description{
This function returns the tabs of my Steam library:
My most played games, my recently played games, and my games with a perfect score.
}
\examples{
remind_me()
}
\keyword{PiPS}
\keyword{Q3.2.1}
\keyword{Steam}
